requirements.txt
astroid==2.2.5
Babel==2.6.0
bcrypt==3.1.6
blinker==1.4
cffi==1.12.2
Click==7.0
colorama==0.4.1
Flask==1.0.2
Flask-AutoIndex==0.6.2
Flask-Babel==0.12.2
Flask-Bcrypt==0.7.1
Flask-Login==0.4.1
Flask-Mail==0.9.1
Flask-Silk==0.2
Flask-SQLAlchemy==2.3.2
Flask-Uploads==0.2.1
Flask-WTF==0.14.2
future==0.17.1
gunicorn==19.9.0
isort==4.3.16
itsdangerous==1.1.0
Jinja2==2.10
lazy-object-proxy==1.3.1
MarkupSafe==1.1.1
mccabe==0.6.1
Pillow==5.4.1
psycopg2-binary==2.8.2
pycparser==2.19
pylint==2.3.1
pytz==2018.9
six==1.12.0
SQLAlchemy==1.3.1
typed-ast==1.3.1
virtualenv==16.4.3
Werkzeug==0.15.1
wrapt==1.11.1
WTForms==2.2.1
